Telegraf

+++ data-processing/excelData.js
@@ -4,8 +4,10 @@ const ExcelJS = require('exceljs');
 
 
 exports.setExcel = (data, switchCheckedObj) => {
-    fs.unlinkSync('./data.xlsx');
-    console.log('file deleted')
+    if(data.date.length != 0) {
+    //     fs.unlinkSync('./data.xlsx');
+    // console.log('file deleted')
+    
     const workbook = new ExcelJS.Workbook();
     const worksheet = workbook.addWorksheet('Data');
 
@@ -54,9 +56,19 @@ exports.setExcel = (data, switchCheckedObj) => {
         }
     }
 
-
-
-    //Save the workbook
-    return workbook.xlsx.writeFile("data.xlsx");
+    try {
+        if (fs.existsSync('./data.xlsx')) {
+            fs.unlinkSync('./data.xlsx');
+            return workbook.xlsx.writeFile("data.xlsx");
+        }
+    } catch {
+        return workbook.xlsx.writeFile("data.xlsx");
+    }
+    // fs.unlinkSync('./data.xlsx');
+    // console.log('file deleted')
+    // //Save the workbook
+    // return workbook.xlsx.writeFile("data.xlsx");
+    }
+    
 
 }
